server/pyproject.toml
[tool.poetry]
name = "flask_security_test_cases"
description=""
version = "1.0.0"
authors = [
    "Martin Millmore <mmillmore@oxfordquantumcircuits.com>",
]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
toml = "^0.10.2"
Flask = "^2.3.2"
flask-cors = "^4.0.0"
SQLAlchemy = "^2.0.12"
psycopg2= "^2.8.6"
flask-security-too = {extras = ["common", "fsqla", "mfa"], version = "^5.3.2"}
python-dotenv = "^1.0.0"
flask-migrate = "^4.0.5"
apispec = "^6.3.0"
apispec-webframeworks = "^0.5.2"
authlib = "^1.2.1"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pytest-dotenv = "^0.5.2"

[tool.isort]
# make it compatible with black
profile = "black" 

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
